+++ manual/source/version_history.rst
@@ -46,7 +46,7 @@ New Features
 * Field maps are now automatically tilted when attached to a tilted beam line element, whereas
   they weren't before.
 * RF cavity fringe fields have been implemented and are on by default. They are controlled with
-  the `includeFringeFields` option.
+  the `includeFringeFieldCavities` option. The `includeFringeFields` option does not affect cavity fringes.
 * Revised executable options for verbosity. These are now the exact same as the intput options. Old
   options are still functional but undocumented.
 * New internal region class allows better setting of defaults when defining custom regions. Preivously,
@@ -60,6 +60,11 @@ New Features
 +------------------------------------+--------------------------------------------------------------------+
 | **Option**                         | **Description**                                                    |
 +====================================+====================================================================+
+| includeFringeFieldCavities         | Include thin fringe fields for RF cavities only, on by default.    |
+|                                    | Cavity fringes are not affected by the includeFringeFields option, |
+|                                    | includeFringeFieldCavities must be explicitly turned off if no     |
+|                                    | fringes are to be built at all in the model.                       |
++------------------------------------+--------------------------------------------------------------------+
 | preprocessGDMLSchema               | Whether to preprocess a copy of the GDML file where the URL of     |
 |                                    | the GDML schema is changed to a local copy provided in BDSIM so    |
 |                                    | geometry can be loaded without internet access. On by default.     |
@@ -175,6 +180,9 @@ General
 Bug Fixes
 ---------
 
+* Fix thick multipole element where the field was 1M times too strong because of the omission of units.
+* Fix Issue #272 where there could be a possible segfault due to the beam particle definition being
+  updated when multiple different particles were used for a `userfile` distribution.
 * Errors in 2D and 3D merged histograms from events were 0 always. The mean was corrected, but the error
   was not filled correctly - this has been fixed.
 * Fix for potential segfault when analysing collimator information branches in event tree. Dependent
@@ -246,6 +254,9 @@ Bug Fixes
   the primary as impacting the wire as the PrimaryFirstHit location.
 * Fixed a bug where the terminator and teleporters would overlap with the tunnel.
 * Fixed two sources of overlaps which may appear when using `lhcleft` or `lhcright` magnet geometries.
+* Placements with respect to thin multipoles would not work. Thin multipoles were always made uniquely
+  where sometimes they didn't have to be - this has been fixed. Also, the searching algorithm has been
+  improved to deal with any uniquely built components, such as rf cavities.
 
 Output Changes
 --------------